Try to work around 'Already in use' error in CI tests

+++ test/lib/whats_opt/openmdao_generator_test.rb
@@ -225,19 +225,33 @@ class OpenmdaoGeneratorTest < ActiveSupport::TestCase
     # assert_match /already been used/, log.join(' ')  # thrift error
   end
 
+  def start_server(dir)
+      pid = spawn("#{WhatsOpt::OpenmdaoGenerator::PYTHON} #{File.join(dir, 'run_server.py')}", [:out] => "/dev/null")
+      sleep(1) # wait 1s for server start
+      pid
+  rescue Errno::EADDRINUSE
+      p "Test server still alive! Just use it!"
+      -1
+  end 
+
+  def stop_server(pid)
+    if pid > 0
+      Process.kill("KILL", pid)
+      Process.waitpid pid;
+    end
+  end
+
   test "should run analysis as default" do
     skip_if_parallel
     skip "Apache Thrift not installed" unless thrift?
     Dir.mktmpdir do |dir|
       @ogen._generate_code dir
-      pid = spawn("#{WhatsOpt::OpenmdaoGenerator::PYTHON} #{File.join(dir, 'run_server.py')}", [:out] => "/dev/null")
-      sleep(1) # wait 1s for server start
+      pid = self.start_server(dir)
       @ogen_remote = WhatsOpt::OpenmdaoGenerator.new(@mda, server_host: "localhost")
       ok, log = @ogen_remote.run
       assert ok
       assert log
-      Process.kill("KILL", pid)
-      Process.waitpid pid; sleep(1)
+      self.stop_server(pid)
     end
   end
 
@@ -246,14 +260,12 @@ class OpenmdaoGeneratorTest < ActiveSupport::TestCase
     skip "Apache Thrift not installed" unless thrift?
     Dir.mktmpdir do |dir|
       @ogen._generate_code dir
-      pid = spawn("#{WhatsOpt::OpenmdaoGenerator::PYTHON} #{File.join(dir, 'run_server.py')}", [:out] => "/dev/null")
-      sleep(1) # wait 1s for server start
+      pid = self.start_server(dir)
       @ogen_remote = WhatsOpt::OpenmdaoGenerator.new(@mda, server_host: "localhost", driver_name: "runonce")
       ok, log = @ogen_remote.run
       assert ok
       assert log
-      Process.kill("KILL", pid)
-      Process.waitpid pid; sleep(1)
+      self.stop_server(pid)
     end
   end
 
@@ -262,8 +274,7 @@ class OpenmdaoGeneratorTest < ActiveSupport::TestCase
     skip "Apache Thrift not installed" unless thrift?
     Dir.mktmpdir do |dir|
       @ogen._generate_code dir
-      pid = spawn("#{WhatsOpt::OpenmdaoGenerator::PYTHON} #{File.join(dir, 'run_server.py')}", [:out] => "/dev/null")
-      sleep(1) # wait 1s for server start
+      pid = self.start_server(dir)
       @ogen_remote = WhatsOpt::OpenmdaoGenerator.new(@mda, server_host: "localhost", driver_name: "smt_doe_lhs")
       File.delete("cicav_doe.sqlite") if File.exist?("cicav_doe.sqlite")
       ok, log = @ogen_remote.run :doe
@@ -271,8 +282,7 @@ class OpenmdaoGeneratorTest < ActiveSupport::TestCase
       assert log
       assert File.exist?("cicav_doe.sqlite")
       File.delete("cicav_doe.sqlite") if File.exist?("cicav_doe.sqlite")
-      Process.kill("KILL", pid)
-      Process.waitpid pid; sleep(1)
+      self.stop_server(pid)
     end
   end
 
@@ -282,8 +292,7 @@ class OpenmdaoGeneratorTest < ActiveSupport::TestCase
     @mda = analyses(:singleton_uq)
     Dir.mktmpdir do |dir|
       @ogen._generate_code dir
-      pid = spawn("#{WhatsOpt::OpenmdaoGenerator::PYTHON} #{File.join(dir, 'run_server.py')}", [:out] => "/dev/null")
-      sleep(1) # wait 1s for server start
+      pid = self.start_server(dir)
       @ogen_remote = WhatsOpt::OpenmdaoGenerator.new(@mda, server_host: "localhost")
       File.delete("singleton_uq_doe.sqlite") if File.exist?("singleton_uq_doe.sqlite")
       ok, log = @ogen_remote.run :doe
@@ -291,8 +300,7 @@ class OpenmdaoGeneratorTest < ActiveSupport::TestCase
       assert log
       assert File.exist?("singleton_uq_doe.sqlite")
       File.delete("singleton_uq_doe.sqlite") if File.exist?("singleton_uq_doe.sqlite")
-      Process.kill("KILL", pid)
-      Process.waitpid pid; sleep(1)
+      self.stop_server(pid)
     end
   end
 
@@ -369,14 +377,12 @@ class OpenmdaoGeneratorTest < ActiveSupport::TestCase
     ogen = WhatsOpt::OpenmdaoGenerator.new(mda)
     Dir.mktmpdir do |dir|
       ogen._generate_code dir
-      pid = spawn("#{WhatsOpt::OpenmdaoGenerator::PYTHON} #{File.join(dir, 'run_server.py')}", [:out] => "/dev/null")
-      sleep(1) # wait 1s for server start
+      pid = self.start_server(dir)
       @ogen_remote = WhatsOpt::OpenmdaoGenerator.new(mda, server_host: "localhost", driver_name: "runonce")
       ok, log = @ogen_remote.run
       assert ok
       assert log
-      Process.kill("KILL", pid)
-      Process.waitpid pid; sleep(1)
+      self.stop_server(pid)
     end
   end
 
@@ -387,14 +393,12 @@ class OpenmdaoGeneratorTest < ActiveSupport::TestCase
     ogen = WhatsOpt::OpenmdaoGenerator.new(mda, pkg_format: true)
     Dir.mktmpdir do |dir|
       ogen._generate_code dir
-      pid = spawn("#{WhatsOpt::OpenmdaoGenerator::PYTHON} #{File.join(dir, 'run_server.py')}", [:out] => "/dev/null")
-      sleep(1) # wait 1s for server start
+      pid = self.start_server(dir)
       @ogen_remote = WhatsOpt::OpenmdaoGenerator.new(mda, pkg_format: true, server_host: "localhost", driver_name: "runonce")
       ok, log = @ogen_remote.run
       assert ok
       assert log
-      Process.kill("KILL", pid)
-      Process.waitpid pid; sleep(1)
+      self.stop_server(pid)
     end
   end
